fix(backup): better logging for backup scripts

+++ scripts/backup-restore.sh
@@ -7,26 +7,36 @@ source "$baseDirectory/ndb-setup/setup.env"
 backupRoot=$BACKUP_DIR
 passphrase=$BACKUP_PASSPHRASE
 
-echo "For which instance do you want to import the backup?"
-read -r org
-
-folder=c-$org
-
 echo "The backup of which day to you want to import? (Format YYYYMMDD e.g. 20220101)"
 read -r date
 
 cd $backupRoot
-echo "$passphrase" | gpg --batch --yes --passphrase-fd 0 -o output -d $date.tar.gz.gpg
-mkdir ~/unpacked
-tar -xzvf output --directory ~/unpacked
 
+echo "decrypting backup ..."
+echo "$passphrase" | gpg --batch --yes --passphrase-fd 0 -o output -d $date.tar.gz.gpg
+
+echo "unpacking backup ..."
+mkdir ~/unpacked
+tar -xzf output --directory ~/unpacked
+
+read -p "Do you want to restore a specific system? [y/n] " choice
+if [[ "$choice" != "y" && "$choice" != "Y" ]]; then
+    echo "Archive unpacked to ~/unpacked. Stopping here."
+    exit 0
+fi
+
+echo "For which instance do you want to import the backup?"
+read -r org
+folder=c-$org
 cd /var/docker/$folder
 docker compose down
 mv couchdb couchdb_tmp
 mv ~/unpacked/var/docker/$folder/couchdb ./couchdb
 docker compose up -d
+echo "backup restored for $folder and instance has been restarted"
 
-#echo "Delete the temporary data?[y/n]"
+
+#echo "Delete the temporary data? [y/n]"
 #read -r delete
 #if [ "$delete" == "y" ] || [ "$delete" == "Y" ]; then
 #    rm -r couchdb_tmp
